requirements.txt
numpy
opencv-python
Pillow
scipy
scikit-image
torch==1.0.1
torchvision==0.2.0
torchfile
tqdm

